Intermediate dev pr (#58)
* Art4 JSON lib removed

* Art4 removed from dependency list

* Autoloader to PSR4. API classes one level up in dir structure. To be used in dev package in packagist

* Vendor name update

* Vendor name update 2

* PSR4 namespace update

* Readme updated with a warning

* Namespace corrections

* Updated the package info with warning

* Namespaces updated for test package

* Composer file update.

* Namespace test

* Namespace test 2

* Namespace test 3

* Readme updated. Now unneeded patreon.php file removed

* Dev keys removed from readme

* Modified the get_data function to allow array, object or raw JSON returns. Renamed __get_data to get_data. get_data function made public so that it is usable directly

* Changed default value of  and added a catch all.

* Changed return_format handling to object var. Set default value. Changed functions to use this var

* Updated the example in readme

* Caching var added to hold the cURL handler. cURL handler function set to return cached handler to bump up performance

* Now deprecated parse parameter removed from API functions

* Reverted curl handler caching since theres no efficient way to separate the suffix. Caching will be used on a per-call basis.

* Reverted recent changes to a state without Art4 lib and with psr4 autoloading for intermediate PR

* Corrected namespace and vendor names for live. Removed test redirect url

* Namespaces in classes corrected

* Indentation styling fix

* Dev comment removed

* Minor styling correction

* Minor readme change

* OAuth.php conflicts resolved. User agent string added to OAuth class cURL requests
