+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.3.2
+
+* Art4 JSON lib removed in preparation for API v2 upgrade
+* API class now allows JSON, array, object formatted returns from the API
+* PSR4 autoloading
+* Directory structure simplified
+* Extra patreon.php file removed
+* Example in readme updated to reflect the changes
+* Unnecessary extra includes removed
+
+## Fixes
+
 # 0.3.1
 
 ## Fixes
