+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "patreon/patreon",
-    "description": "Interact with the Patreon API via OAuth",
+    "description": "Interact with the Patreon API via OAuth.",
     "license": "MIT",
     "authors": [
         {
@@ -10,12 +10,11 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=5.3.0",
-        "art4/json-api-client": "0.9.*"
+        "php": ">=5.3.0"
     },
     "autoload": {
-        "psr-0": {
-            "Patreon": "src/"
+        "psr-4": {
+            "Patreon\\": "src/"
         }
     }
 }
