--- src/patreon.php
@@ -1,5 +1,0 @@
-<?php
-
-require_once(__DIR__ . '/../vendor/autoload.php');
-require_once("Patreon/API.php");
-require_once("Patreon/OAuth.php");
